Compile OSX, new version ACE,OpenSSL

+++ src/game/Maps/MapPersistentStateMgr.cpp
@@ -794,7 +794,7 @@ void MapPersistentStateManager::PackInstances()
         delete result;
     }
 
-    BarGoLink bar(InstanceSet.size() + 1);
+    BarGoLink bar((uint32)(InstanceSet.size() + 1));
     bar.step();
 
     uint32 InstanceNumber = RESERVED_INSTANCES_LAST;
